requirements.txt
aiohttp
discord.py
beautifulsoup4